Address remaining review comments
- Fix typeguard to not use `as` (use `in` operator instead)
- Move email reply logic to lib/api/assistant/email_reply.ts
- Keep thin activity wrappers in temporal/agent_loop/activities/
